Update project with recent changes
- Add masking schedule analysis and test scripts
- Add MNIST diffusion tests
- Update KITTI dataset and world model training
- Add AGENTS.md for AI assistant guidance
- Add comparison script for training objectives
- Clean up old server setup scripts
- Update .gitignore to exclude .npy files

+++ copilot4d/data/kitti_sequence_dataset.py
@@ -125,7 +125,7 @@ class KITTISequenceDataset(Dataset):
             tokens = torch.load(token_file, map_location="cpu", weights_only=True)
             tokens_list.append(tokens)
         
-        tokens = torch.stack(tokens_list, dim=0)  # (T, H, W)
+        tokens = torch.stack(tokens_list, dim=0).long()  # (T, H, W)
         
         # Load poses
         with open(seq_path / "poses.pkl", "rb") as f:
